requirements.txt
python-dotenv
backoff
pygame
gql
